+++ outputs.tf
@@ -8,7 +8,7 @@ output "kms_key_id" {
   value       = module.kms_key.id
 }
 
-output "monitor_iam_access_sns_topic_arn" {
-  description = "ARN of the SNS Topic in the Audit account for IAM access monitoring notifications"
-  value       = aws_sns_topic.monitor_iam_access_audit.arn
+output "monitor_iam_access_event_bus_arn" {
+  description = "ARN of the Event Bus in the Audit account for IAM access monitoring notifications"
+  value       = aws_cloudwatch_event_bus.monitor_iam_access_audit.arn
 }